moved renumbering to upper app level
